feat: add kaizen introduction and guest routes
